20220520_airflow_vscode/requirements.txt
apache-airflow-providers-amazon == 3.0.0
boto3 == 1.18.65
botocore == 1.21.65
